doing automation

+++ mongodb.sh
@@ -2,17 +2,17 @@
 CHECK(){
     if [ $? -ne 0 ] 
     then
-        echo "$1 not installed successfully"
+        echo "$1 not $R installed$N successfully"
         exit 1
     else
-        echo "$1 installed successfully"
+        echo "$1 $Ginstalled$N successfully"
     fi
 }
 ID=$(id -u)
 R="\e[31m"
 G="\e[32m"
 Y="\e[33m"
-NORMAL="\e[0m"
+N="\e[0m"
 
 if [ "$ID" -ne 0 ];
 then
@@ -21,6 +21,20 @@ else
     echo -e "$G ROOT USER"  
 fi
 
-cp mongo.txt /etc/yum.repos.d/mongo.repo
+if [ ! -e "/etc/yum.repos.d/mongo.repo" ]; then
+    cp mongo.txt /etc/yum.repos.d/mongo.repo
+fi
 CHECK "copy"
 
+dnf install mongodb-org -y 
+CHECK "mongodb"
+
+systemctl enable mongod
+CHECK "enable"
+
+systemctl start mongod
+CHECK "start"
+sed -i 's/127\.0\.0\.1/0\.0\.0\.0/' /etc/mongod.conf
+systemctl restart mongod
+CHECK "Restart"
+echo "Mongodb script runned successfully"
